requirements.txt
openai 
streamlit==1.28.0
pandas==2.1.1
plotly==5.17.0
pillow==10.0.1